refactor(ws): extract adopt_codex_identity shared adoption tail from the candidate channel
